requirements.txt
# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
networkx>=2.8.0

# Jupyter Notebook
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.0.0

# Date/Time handling
python-dateutil>=2.8.0

# For advanced forecasting
statsmodels>=0.14.0
scikit-learn>=1.3.0
prophet>=1.1.0
xgboost>=2.0.0
lightgbm>=4.0.0

# Web Scraping and NLP
newspaper4k
playwright
transformers
torch
lxml_html_clean
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Web Server (for news fetch API)
flask>=2.3.0
flask-cors>=4.0.0
feedparser